backend/requirements.txt
alembic==1.11.1
anyio==3.7.0
asyncpg==0.27.0
beanie==1.19.2
cffi==1.15.1
click==8.1.3
cryptography==40.0.2
dnspython==2.3.0
ecdsa==0.18.0
exceptiongroup==1.1.1
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
lazy-model==0.0.5
Mako==1.2.4
MarkupSafe==2.1.2
motor==3.1.2
psycopg2-binary==2.9.6
pyasn1==0.5.0
pycparser==2.21
pydantic==1.10.8
pymongo==4.3.3
python-dotenv==1.0.0
python-jose==3.3.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
toml==0.10.2
typing_extensions==4.6.2
uvicorn==0.22.0
